Split received/case.py and received/actor.py into subpackages (#881)
Convert two large use-case modules (819 and 803 lines) into focused
subpackages, each under 500 lines per submodule (CS-18-002).

received/case/ submodules:
- _helpers.py: private helper functions shared across the package
- create.py: CreateCaseReceivedUseCase (bootstrap path)
- update.py: UpdateCaseReceivedUseCase
- engage_defer.py: EngageCaseReceivedUseCase, DeferCaseReceivedUseCase
- lifecycle.py: AddReportToCaseReceivedUseCase, CloseCaseReceivedUseCase
- validate.py: InvalidateCaseUseCase, CloseCaseUseCase, ValidateCaseUseCase
- __init__.py: re-exports all public names (AC-3)

received/actor/ submodules:
- suggest.py: Suggest/Accept/Reject SuggestActorToCase
- case_manager_role.py: Offer/Accept/Reject CaseManagerRole
- ownership.py: Offer/Accept/Reject CaseOwnershipTransfer
- invite.py: Invite/Accept/Reject InviteActorToCase
- announce.py: AnnounceVulnerabilityCaseReceivedUseCase
- __init__.py: re-exports all public names including _find_case_actor_id

Test layout mirrors source split (AC-4):
- test/core/use_cases/received/case/ (test_create, test_update, test_engage_defer)
- test/core/use_cases/received/actor/ (test_invite, test_suggest, test_ownership, test_case_manager_role, test_announce)

All 3205 unit tests pass. Black, flake8, mypy, pyright clean.
